+++ react-components/src/apps.js
@@ -1,5 +1,6 @@
 import * as details from "./apps/details";
 import * as admin from "./apps/admin";
 import * as quickLaunch from "./apps/quickLaunch";
+import * as listing from "./apps/listing";
 
-export { details, admin, quickLaunch };
+export { details, admin, quickLaunch, listing };
