+++ react-components/src/apps/listing/ids.js
@@ -1,7 +1,16 @@
 export default {
-    TABLE_HEADER: {
+    FIELD: {
         NAME: "name",
         INTEGRATED_BY: "integratedBy",
         SYSTEM: "system",
+        RATING: "rating",
+    },
+    LISTING: {
+        TYPE: "type",
+        AGAVE: ".agave",
+        DE: ".de",
+        INTERACTIVE: ".interactive",
+        OSG: ".osg",
+        SORT: ".sort",
     },
 };
