+++ react-components/src/apps/style.js
@@ -2,6 +2,7 @@ import fav from "../resources/images/fav.png";
 import favAdd from "../resources/images/fav_add.png";
 import favRemove from "../resources/images/fav_remove.png";
 import notFav from "../resources/images/not_fav.png";
+import { palette } from "@cyverse-de/ui-lib";
 
 export default (theme) => ({
     statFilterButton: {
